% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.r
\name{calc_zeta_discrete}
\alias{calc_zeta_discrete}
\title{Compute variance inflation factor from uncertainty in responding sample size
(for integer \eqn{n_h})}
\usage{
calc_zeta_discrete(
  n_h,
  phibar_h,
  rh_min = 3.5,
  round_flag = TRUE,
  verbose_flag = FALSE
)
}
\arguments{
\item{n_h}{(vector) strata sample sizes before nonresponse (\eqn{n_h})}

\item{phibar_h}{(vector) strata response propensities (\eqn{\bar{\phi}_h})}

\item{rh_min}{(scalar) minimum target respondents per stratum (\eqn{r_h^{LB}}); default is 3.5}

\item{round_flag}{(bool) specifies whether to round continuous allocations
(versus throwing an error)}

\item{verbose_flag}{(bool) flag on whether to provide detailed results}
}
\value{
vector of length \eqn{H} containing
\eqn{\left\{\zeta_h(n_h', \bar{\phi}_h):h=1,2,...,H\right\}},
where \eqn{n_h'} is the larger of \eqn{n_h} or \eqn{\frac{r_h^{LB}}{\bar{\phi}_h}}
}
\description{
Computes
\eqn{\zeta_h(n_h, \bar{\phi}_h):=\mathrm{E}(r_h)\mathrm{E}\left(\frac{1}{r_h}\right)},
as defined in the paper.
If there are any strata where the allocation may lead to
\eqn{\mathrm{E}(r_h) < r_h^{LB}} for user-specified \eqn{r_h^{LB}}
(3.5 by default), then \eqn{\zeta_h(.)} is evaluated
using \eqn{n_h':= max\left(n_h, \left\lceil
     \frac{r_h^{LB}}{\bar{\phi}_h}\right\rceil\right)}
in place of \eqn{n_h}.
Continuous values of \eqn{n_h} are rounded (by default).
}
\details{
In our paper, we assumed that the number of respondents in stratum \eqn{h}
can be modeled as standard binomial with support for zero removed
(i.e., zero-truncated binomial; see \code{\link[=dtruncbinom]{dtruncbinom()}}),
written as \eqn{r_h \sim TBinom(n_h, \bar{\phi}_h)}, where
\eqn{n_h} is the number of invitees in stratum \eqn{h},
\eqn{\bar{\phi}_h} is the average response propensity within stratum \eqn{h},
and where the unit-level response propensities are assumed constant within
strata.
Our paper defines the function
\eqn{\zeta_h(n_h, \bar{\phi}_h):=\mathrm{E}(r_h)\mathrm{E}\left(\frac{1}{r_h}\right)}.
This quantity is a variance inflation term that captures the effect of
variability in the number of respondents for a given allocation.

The current function (\code{calc_zeta_discrete}())
computes \eqn{\zeta_h(n_h', \bar{\phi}_h)}, where we use
\eqn{n_h':= max\left(n_h, \left\lceil \frac{r_h^{LB}}{\bar{\phi}_h}\right\rceil\right)}
to avoid underallocating to strata with too few expected respondents, and
where \eqn{r_h^{LB}} is some given lower bound on the number
of expected respondents.
By default, we set \eqn{r_h^{LB} = 3.5}, since the truncated binomial
distribution may sometimes be a poor approximation
for the binomial distribution below these levels,
and as we observed numerically that \eqn{\zeta_h(n_h, \bar{\phi}_h)}
is roughly maximized for various \eqn{\bar{\phi}_h} (fixed at levels
between .01 and 1) when \eqn{n_h \approx \frac{3.5}{\bar{\phi}_h}}.

Function is only defined for discrete \eqn{n_h}; input \code{round_flag}
controls whether \eqn{n_h} is rounded to the nearest integer
(as opposed to throwing an error).
}
\examples{
\dontrun{
#Basic example
calc_zeta_discrete(n_h = c(100, 200, 300, 300),
                  phibar_h = c(.03, .02, .05, .001))

#Verbose info shows that strata 1,5 have their sample sizes increased
#   due to having fewer than 3.5 expected respondents
calc_zeta_discrete(n_h = c(100, 200, 300, 300),
                   phibar_h = c(.03, .02, .05, .001),
                   verbose_flag = TRUE)

#Similar to above, but removes the minimum number of respondents,
# which changes the values for strata 1 and 4.
calc_zeta_discrete(n_h = c(100, 200, 300, 300),
                  phibar_h = c(.03, .02, .05, .001),
                  rh_min = 0,
                  verbose_flag = TRUE)
}
}
\seealso{
\code{\link[=calc_zeta]{calc_zeta()}} for use with continuous \code{n_h}
}
\keyword{distributions}
\keyword{internal}
