% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingNR-package.R
\docType{package}
\name{samplingNR}
\alias{samplingNR}
\alias{samplingNR-package}
\title{samplingNR: A package for stratified sample allocation under anticipated nonresponse}
\description{
Package contains functions used for computing stratified sample allocations
in a manner that accounts for the anticipated effects of nonresponse
on sample efficiency.
}
\details{
To get started, try running \code{vignette("samplingNR")} or view
the help page for \code{\link[=opt_nh_nonresp]{opt_nh_nonresp()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmendelson256/samplingNR}
  \item Report bugs at \url{https://github.com/jmendelson256/samplingNR/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Mendelson \email{jmendelson@gmail.com} (\href{https://orcid.org/0000-0001-5575-8773}{ORCID})

}
\keyword{internal}
